backend/pyproject.toml
[tool.poetry]
name = "backend"
version = "0.1.0"
description = "Backend for a dynamic portfolio website showcasing IT services for ITMANO"

[tool.poetry.dependencies]
python = "^3.9"
django = "^4.2.0"
djangorestframework = "^3.14.0"
psycopg2-binary = "^2.9.5"
django-cors-headers = "^4.0.0"
python-memcached = "^1.59"
django-environ = "^0.10.0"
django-storages = "^1.13.2"
gunicorn = "^21.0.0"
pymemcache = "^4.0.0"
pillow = "^11.2.1"
environ = "^1.0"
celery = "^5.5.2"
redis = "^6.0.0"

[tool.poetry.group.dev.dependencies]
black = "^23.3.0"
isort = "^5.12.0"
flake8 = "^6.0.0"
pytest = "^7.3.1"
pytest-django = "^4.5.2"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
